requirements.txt
h5py==3.9.0
matplotlib==3.7.2
numpy==1.25.0
Pillow==10.0.0
scipy==1.11.1
torch==2.0.1
torchvision==0.15.2
opencv-python==4.10.0.84
streamlink==6.10.0
ultralytics==8.2.94
tensorflow==2.16.1
torch==2.4.1
torchvision==0.19.1